Publish DEPS for Chromium 30.0.1599.120
Incrementing VERSION to 30.0.1599.120

git-svn-id: svn://svn.chromium.org/chrome/branches/1599/src@235169 0039d316-1c4b-4281-b951-d872f2087c98

+++ tools/telemetry/telemetry/core/chrome/desktop_browser_backend.py
@@ -9,17 +9,19 @@ import sys
 import tempfile
 
 from telemetry.core import util
-from telemetry.core.chrome import browser_backend
-from telemetry.core.chrome import cros_util
+from telemetry.core.backends import browser_backend
+from telemetry.core.backends.chrome import chrome_browser_backend
 
-class DesktopBrowserBackend(browser_backend.BrowserBackend):
+class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
   """The backend for controlling a locally-executed browser instance, on Linux,
   Mac or Windows.
   """
-  def __init__(self, options, executable, is_content_shell):
+  def __init__(self, options, executable, flash_path, is_content_shell,
+               browser_directory, delete_profile_dir_after_run=True):
     super(DesktopBrowserBackend, self).__init__(
         is_content_shell=is_content_shell,
-        supports_extensions=not is_content_shell, options=options)
+        supports_extensions=not is_content_shell,
+        options=options)
 
     # Initialize fields so that an explosion during init doesn't break in Close.
     self._proc = None
@@ -30,28 +32,40 @@ class DesktopBrowserBackend(browser_backend.BrowserBackend):
     if not self._executable:
       raise Exception('Cannot create browser, no executable found!')
 
+    self._flash_path = flash_path
+    if self._flash_path and not os.path.exists(self._flash_path):
+      logging.warning(('Could not find flash at %s. Running without flash.\n\n'
+                       'To fix this see http://go/read-src-internal') %
+                      self._flash_path)
+      self._flash_path = None
+
     if len(options.extensions_to_load) > 0 and is_content_shell:
       raise browser_backend.ExtensionsNotSupportedException(
           'Content shell does not support extensions.')
 
+    self._browser_directory = browser_directory
     self._port = util.GetAvailableLocalPort()
+    self._profile_dir = None
     self._supports_net_benchmarking = True
-    self._LaunchBrowser(options)
+    self._delete_profile_dir_after_run = delete_profile_dir_after_run
 
-    # For old chrome versions, might have to relaunch to have the
-    # correct benchmarking switch.
-    if self._chrome_branch_number < 1418:
-      self.Close()
-      self._supports_net_benchmarking = False
-      self._LaunchBrowser(options)
+    self._SetupProfile()
 
-    if self.options.cros_desktop:
-      cros_util.NavigateLogin(self)
+  def _SetupProfile(self):
+    if not self.options.dont_override_profile:
+      self._tmpdir = tempfile.mkdtemp()
+      profile_dir = self._profile_dir or self.options.profile_dir
+      if profile_dir:
+        if self.is_content_shell:
+          logging.critical('Profiles cannot be used with content shell')
+          sys.exit(1)
+        shutil.rmtree(self._tmpdir)
+        shutil.copytree(profile_dir, self._tmpdir)
 
-  def _LaunchBrowser(self, options):
+  def _LaunchBrowser(self):
     args = [self._executable]
     args.extend(self.GetBrowserStartupArgs())
-    if not options.show_stdout:
+    if not self.options.show_stdout:
       self._tmp_output_file = tempfile.NamedTemporaryFile('w', 0)
       self._proc = subprocess.Popen(
           args, stdout=self._tmp_output_file, stderr=subprocess.STDOUT)
@@ -70,31 +84,49 @@ class DesktopBrowserBackend(browser_backend.BrowserBackend):
     args.append('--remote-debugging-port=%i' % self._port)
     if not self.is_content_shell:
       args.append('--window-size=1280,1024')
+      if self._flash_path:
+        args.append('--ppapi-flash-path=%s' % self._flash_path)
       if self._supports_net_benchmarking:
         args.append('--enable-net-benchmarking')
       else:
         args.append('--enable-benchmarking')
       if not self.options.dont_override_profile:
-        self._tmpdir = tempfile.mkdtemp()
-        if self.options.profile_dir:
-          if self.is_content_shell:
-            logging.critical('Profiles cannot be used with content shell')
-            sys.exit(1)
-          shutil.rmtree(self._tmpdir)
-          shutil.copytree(self.options.profile_dir, self._tmpdir)
         args.append('--user-data-dir=%s' % self._tmpdir)
-      if self.options.cros_desktop:
-        ext_path = os.path.join(os.path.dirname(__file__), 'chromeos_login_ext')
-        args.extend(['--login-manager', '--login-profile=user',
-                     '--stub-cros', '--login-screen=login',
-                     '--auth-ext-path=%s' % ext_path])
     return args
+
+  def SetProfileDirectory(self, profile_dir):
+    # Make sure _profile_dir hasn't already been set.
+    assert self._profile_dir is None
+
+    if self.is_content_shell:
+      logging.critical('Profile creation cannot be used with content shell')
+      sys.exit(1)
+
+    self._profile_dir = profile_dir
+
+  def Start(self):
+    self._LaunchBrowser()
+
+    # For old chrome versions, might have to relaunch to have the
+    # correct net_benchmarking switch.
+    if self._chrome_branch_number < 1418:
+      self.Close()
+      self._supports_net_benchmarking = False
+      self._LaunchBrowser()
 
   @property
   def pid(self):
     if self._proc:
       return self._proc.pid
     return None
+
+  @property
+  def browser_directory(self):
+    return self._browser_directory
+
+  @property
+  def profile_directory(self):
+    return self._tmpdir
 
   def IsBrowserRunning(self):
     return self._proc.poll() == None
@@ -107,6 +139,12 @@ class DesktopBrowserBackend(browser_backend.BrowserBackend):
         return f.read()
     except IOError:
       return ''
+
+  def GetStackTrace(self):
+    # crbug.com/223572, symbolize stack trace for desktop browsers.
+    logging.warning('Stack traces not supported on desktop browsers, '
+                    'returning stdout')
+    return self.GetStandardOutput()
 
   def __del__(self):
     self.Close()
@@ -139,7 +177,8 @@ class DesktopBrowserBackend(browser_backend.BrowserBackend):
           self._proc = None
           raise Exception('Could not shutdown the browser.')
 
-    if self._tmpdir and os.path.exists(self._tmpdir):
+    if self._delete_profile_dir_after_run and \
+        self._tmpdir and os.path.exists(self._tmpdir):
       shutil.rmtree(self._tmpdir, ignore_errors=True)
       self._tmpdir = None
 
@@ -148,17 +187,4 @@ class DesktopBrowserBackend(browser_backend.BrowserBackend):
       self._tmp_output_file = None
 
   def CreateForwarder(self, *port_pairs):
-    return DoNothingForwarder(*port_pairs)
-
-
-class DoNothingForwarder(object):
-  def __init__(self, *port_pairs):
-    self._host_port = port_pairs[0].local_port
-
-  @property
-  def url(self):
-    assert self._host_port
-    return 'http://127.0.0.1:%i' % self._host_port
-
-  def Close(self):
-    self._host_port = None
+    return browser_backend.DoNothingForwarder(*port_pairs)
